feat: add LexicographicWord
