+++ miden-crypto/src/lib.rs
@@ -22,9 +22,6 @@ pub use winter_math::{
 };
 pub use word::{Word, WordError};
 
-// TYPE ALIASES
-// ================================================================================================
-
 // CONSTANTS
 // ================================================================================================
